+++ Kogel.Slave.Mysql/Events/RotateEvent.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Buffers;
 using System.Text;
-using Kogel.Slave.Mysql.Extension;
 using SuperSocket.ProtoBase;
 
 namespace Kogel.Slave.Mysql
@@ -17,7 +16,7 @@ namespace Kogel.Slave.Mysql
 
         protected internal override void DecodeBody(ref SequenceReader<byte> reader, object context)
         {
-            reader.TryReadLittleEndian(out short position);
+            reader.TryReadLittleEndian(out long position);
             RotatePosition = position;
 
             var binglogFileNameSize = reader.Remaining - (int)LogEvent.ChecksumType;
