+++ Kogel.Slave.Mysql/Events/XIDEvent.cs
@@ -11,7 +11,7 @@ namespace Kogel.Slave.Mysql
         
         protected internal override void DecodeBody(ref SequenceReader<byte> reader, object context)
         {
-            reader.TryReadLittleEndian(out short tarnsID);
+            reader.TryReadLittleEndian(out long tarnsID);
             TransactionID = tarnsID;
         }
     }
